# docs/openapi.json
{
  "components": {
    "schemas": {
      "AvailabilityRequest": {
        "properties": {
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "state"
        ],
        "title": "AvailabilityRequest",
        "type": "object"
      },
...
      "CreateOrderRequest": {
        "properties": {
          "customerPhone": {
            "anyOf": [
              {
                "maxLength": 24,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customerphone"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderItemRequest"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Items",
            "type": "array"
          },
          "note": {
            "default": "",
            "maxLength": 500,
            "title": "Note",
            "type": "string"
          },
          "storeId": {
            "maxLength": 36,
            "minLength": 1,
            "title": "Storeid",
            "type": "string"
          }
        },
        "required": [
          "storeId",
          "items"
        ],
        "title": "CreateOrderRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MerchantMeOut": {
        "properties": {
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "memberships": {
            "items": {
              "$ref": "#/components/schemas/MerchantMembershipOut"
            },
            "title": "Memberships",
            "type": "array"
          },
          "userId": {
            "title": "Userid",
            "type": "string"
          }
        },
        "required": [
          "userId",
          "displayName",
          "memberships"
        ],
        "title": "MerchantMeOut",
        "type": "object"
      },
...
      "MerchantOrderListItem": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "itemCount": {
            "title": "Itemcount",
            "type": "integer"
          },
          "publicOrderCode": {
            "title": "Publicordercode",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "storeId": {
            "title": "Storeid",
            "type": "string"
          },
          "totalAmount": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Totalamount",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "publicOrderCode",
          "storeId",
          "state",
          "itemCount",
          "totalAmount",
          "currency",
          "version",
          "createdAt"
        ],
        "title": "MerchantOrderListItem",
        "type": "object"
      },
...
      "PriceCreateRequest": {
        "properties": {
          "amount": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 999999999.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Amount"
          },
          "currency": {
            "maxLength": 8,
            "minLength": 1,
            "title": "Currency",
            "type": "string"
          }
        },
        "required": [
          "amount",
          "currency"
        ],
        "title": "PriceCreateRequest",
        "type": "object"
      },
      "ProductCreateRequest": {
        "properties": {
          "category": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Category",
            "type": "string"
          },
          "categoryAr": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Categoryar"
          },
          "description": {
            "default": "",
            "maxLength": 2000,
            "title": "Description",
            "type": "string"
          },
          "descriptionAr": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descriptionar"
          },
          "imageRef": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imageref"
          },
          "name": {
            "maxLength": 140,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "nameAr": {
            "anyOf": [
              {
                "maxLength": 140,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Namear"
          },
          "sku": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Sku",
            "type": "string"
          },
          "storeId": {
            "title": "Storeid",
            "type": "string"
          },
          "unitLabel": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Unitlabel",
            "type": "string"
          }
        },
        "required": [
          "storeId",
          "sku",
          "name",
          "category",
          "unitLabel"
        ],
        "title": "ProductCreateRequest",
        "type": "object"
      },
      "ProductPatchRequest": {
        "properties": {
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          },
          "category": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "categoryAr": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Categoryar"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 2000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "descriptionAr": {
            "anyOf": [
              {
                "maxLength": 2000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descriptionar"
          },
          "imageRef": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imageref"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 140,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "nameAr": {
            "anyOf": [
              {
                "maxLength": 140,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Namear"
          },
          "unitLabel": {
            "anyOf": [
              {
                "maxLength": 40,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unitlabel"
          }
        },
        "title": "ProductPatchRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Offline-first Click & Collect portfolio API.",
    "title": "Jahhezly API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/merchant/catalog/products": {
      "post": {
        "operationId": "add_product_v1_merchant_catalog_products_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Add Product V1 Merchant Catalog Products Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Product",
        "tags": [
          "merchant"
        ]
      }
    },
    "/v1/merchant/catalog/products/{product_id}": {
      "patch": {
        "operationId": "update_product_v1_merchant_catalog_products__product_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "title": "Product Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Update Product V1 Merchant Catalog Products  Product Id  Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Product",
        "tags": [
          "merchant"
        ]
      }
    },
    "/v1/merchant/catalog/products/{product_id}/availability": {
      "post": {
        "operationId": "availability_v1_merchant_catalog_products__product_id__availability_post",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "title": "Product Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AvailabilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Availability V1 Merchant Catalog Products  Product Id  Availability Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Availability",
        "tags": [
          "merchant"
        ]
      }
    },
    "/v1/merchant/catalog/products/{product_id}/price": {
      "post": {
        "operationId": "add_price_v1_merchant_catalog_products__product_id__price_post",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "title": "Product Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PriceCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Add Price V1 Merchant Catalog Products  Product Id  Price Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Price",
        "tags": [
          "merchant"
        ]
      }
    },
    "/v1/merchant/me": {
      "get": {
        "operationId": "me_v1_merchant_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantMeOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Me",
        "tags": [
          "merchant"
        ]
      }
    },
    "/v1/merchant/orders": {
      "get": {
        "operationId": "list_orders_v1_merchant_orders_get",
        "parameters": [
          {
            "in": "query",
            "name": "store_id",
            "required": true,
            "schema": {
              "title": "Store Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MerchantOrderListItem"
                  },
                  "title": "Response List Orders V1 Merchant Orders Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Orders",
        "tags": [
          "merchant"
        ]
      }
    },
    "/v1/merchant/orders/{order_id}": {
      "get": {
        "operationId": "get_order_v1_merchant_orders__order_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Order",
        "tags": [
          "merchant"
        ]
      }
    },
    "/v1/merchant/orders/{order_id}/accept": {
      "post": {
        "operationId": "accept_order_v1_merchant_orders__order_id__accept_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "title": "Order Id"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Accept Order",
        "tags": [
          "merchant"
        ]
      }
    },
    "/v1/merchant/orders/{order_id}/collect": {
      "post": {
        "operationId": "collect_order_v1_merchant_orders__order_id__collect_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "title": "Order Id"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Collect Order",
        "tags": [
          "merchant"
        ]
      }
    },
...
    "/v1/merchant/orders/{order_id}/reject": {
      "post": {
        "operationId": "reject_order_v1_merchant_orders__order_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "title": "Order Id"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reject Order",
        "tags": [
          "merchant"
        ]
      }
    },
    "/v1/merchant/orders/{order_id}/start-preparing": {
      "post": {
        "operationId": "preparing_order_v1_merchant_orders__order_id__start_preparing_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "title": "Order Id"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Preparing Order",
        "tags": [
          "merchant"
        ]
      }
    },
    "/v1/orders": {
      "post": {
        "operationId": "create_order_v1_orders_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "X-Guest-Session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Guest-Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Order",
        "tags": [
          "customer"
        ]
      }
    },
    "/v1/orders/{order_id}": {
      "get": {
        "operationId": "get_order_v1_orders__order_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "title": "Order Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Guest-Session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Guest-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Order",
        "tags": [
          "customer"
        ]
      }
    },
...
  }
}